+++ src/components/Checkbox/composables/index.ts
@@ -1,6 +1,6 @@
 import { computed } from 'vue'
 import type { Ref } from 'vue'
-import { useGlobalHiddenInputClasses } from '../../../composables/index'
+import { useGlobalHiddenInputClasses } from '../../../composables'
 import type { Description, Rounded, Size, State } from '@/globals/types'
 import type { TwClassInterface } from '@/globals/interfaces'
