+++ src/components/Modal/LuiModal.vue
@@ -10,7 +10,7 @@ import { computed, watch } from 'vue'
 import type { PropType } from 'vue'
 import { UseFocusTrap } from '@vueuse/integrations/useFocusTrap/component'
 import { useId } from '../../utils/useId'
-import { useTeleportWrapper } from '../../composables/useTeleportWrapper'
+import { useTeleportWrapper } from '../../composables'
 import LuiButton from '../Button/LuiButton.vue'
 import type { TwClassInterface } from '@/globals/interfaces'
 import type { Size } from '@/globals/types'
