+++ src/components/MultiSelect/LuiMultiSelect.vue
@@ -14,10 +14,8 @@ import type { TwClassInterface } from '../../globals/interfaces'
 import { ContextKey } from '../Select/symbols'
 
 // import { useProperPosition } from '../../composables/useProperPosition'
-import { useOutsideClick } from '../../composables/useOutsideClick'
+import { useMenuStyles, useOutsideClick, useTeleportWrapper } from '../../composables'
 import { useInputClasses } from '../Input/composables/index'
-import { useTeleportWrapper } from '../../composables/useTeleportWrapper'
-import { useMenuStyles } from '../../composables/useMenuStyles'
 import { useId } from '../../utils/useId'
 import { isElementScrollable } from '../../utils/isElementScrollable'
 import { hasSlotContent } from '../../utils/hasSlotContent'
